PDJB-1168: split out org and single landlord cyas
